fix css bug in expense overview budget profit and add sweet alert for category modal

+++ src/app/components/common/category-modal/category-modal.ts
@@ -157,11 +157,30 @@ export class CategoryModalComponent implements OnInit {
 
   async onDelete(categoryId: string): Promise<void> {
     this.deletingStates[categoryId] = true;
+    const confirmMsg = await firstValueFrom(
+      this.translateService.get('CONFIRM_DELETE_CATEGORY')
+    );
     try {
-      await this.categoryService.deleteCategory(categoryId);
-      await this.loadCategories();
-      this.categoryAdded.emit();
-      Toast.fire({ icon: 'success', title: this.translateService.instant('CATEGORY_DELETED_SUCCESS') });
+      Swal.fire({
+        title: this.translateService.instant('CONFIRM_DELETE_TITLE'),
+        text: confirmMsg,
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: this.translateService.instant('DELETE_BUTTON'),
+        cancelButtonText: this.translateService.instant('CANCEL_BUTTON'),
+        reverseButtons: true
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          try {
+            await this.categoryService.deleteCategory(categoryId);
+            await this.loadCategories();
+            this.categoryAdded.emit();
+            Toast.fire({ icon: 'success', title: this.translateService.instant('CATEGORY_DELETED_SUCCESS') });      
+          } catch (error: any) {
+            Toast.fire({ icon: 'error', title: this.translateService.instant('DATA_DELETE_ERROR') }); 
+          }
+        }
+      });
     } catch (error) {
       console.error(`Error deleting category ${categoryId}:`, error);
     } finally {
